rewrite making it more functional

+++ src/main/scala/de/htwg/se/blackjackKN/model/betComponent/Bet.scala
@@ -1,4 +1,4 @@
 package de.htwg.se.blackjackKN.model.betComponent
 
 
-case class Bet(value: Double = 100) {}
+case class Bet(value: Int = 100) {}
